--- a/openh264/Cargo.toml
+++ b/openh264/Cargo.toml
@@ -22,6 +22,7 @@ backtrace = [] # Remove once backtrace feature is stable.
 
 [dependencies]
 openh264-sys2 = { path = "../openh264-sys2", version = "0.6.0", default-features = false }
+rayon = "1.10.0"
 
 [package.metadata.docs.rs]
 all-features = true