@ralfbiedert 2024-11-21
Thanks for the PR. Unfortunately [we've had that PR before](https://github.com/ralfbiedert/openh264-rs/pull/27/files), also [see my comment there](https://github.com/ralfbiedert/openh264-rs/pull/27#issuecomment-1370800587). In brief:

- rayon is good for apps, not for libs used in controlled "low latency high performance" contexts (e.g., game engines)
- the speedup is poor for that it probably uses 4x, 8x or 16x cores and spins up an entire thread pool
- if you really wanted "thread parallelism" there should be a way for the API to expose "pixel chunks" a caller could process in parallel (this is a bit tricky to design, but I'd argue it's the way to go for threads)
- a much better parallelization would be SIMD and / or autovectorization