@jnnks 2024-11-21
Ok, I will look into SIMD.